Refactor: Update imports for user service and types in RoleBasedRedirect, ThreadDetail, and ThreadList components
